pyproject.toml
[project]
name = "langchain-the-framework"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "ddgs>=9.10.0",
    "langchain>=1.2.9",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.9",
    "openai>=2.21.0",
    "python-dotenv>=1.2.1",
    "wikipedia>=1.4.0",
]
